Return nil instead of Place object in case of no results
This tells the user of fetch() method that there is no results,
so that case when there is no reverse geocode results can be handled.

Incidentally, this avoids a TypeError,
“TypeError: no implicit conversion of Symbol into Integer”,
which is caused by Place#address(), as @attrs will be an array,
as Nominatim returns array as response when there is no results.